successfully extract type system to a separate file
