+++ bigraph_schema/type_system_tests.py
@@ -7,9 +7,10 @@ import pprint
 import numpy as np
 from dataclasses import  asdict
 
-from bigraph_schema.type_system import (
-    TypeSystem, divide_longest, base_type_library, accumulate, to_string, deserialize_integer, apply_schema, diff
+from bigraph_schema.type_functions import (
+    divide_longest, base_type_library, accumulate, to_string, deserialize_integer, apply_schema, diff
 )
+from bigraph_schema import TypeSystem
 from bigraph_schema.units import units
 from bigraph_schema.registry import establish_path, remove_omitted, NONE_SYMBOL
 import bigraph_schema.data as data
